backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.25
pydantic==1.10.18
python-dotenv==1.0.1
apscheduler==3.10.4
loguru==0.7.2
requests==2.32.3
pandas==2.2.3
tushare==1.4.24
python-multipart==0.0.9
psycopg2-binary==2.9.9

